[Modal content] Code refactoring
